requirements.txt
flask==3.1.0
flask-session==0.8.0
yt_dlp==2025.3.31
requests==2.32.3
python-dotenv==0.21.0
langchain==0.3.24
langchain-community==0.3.23
langchain-openai==0.3.16
openai==1.68.2
youtube-search==2.1.2